requirements.txt
python-telegram-bot
gspread
pdfplumber
oauth2client
gunicorn
